requirements.txt
streamlit
sounddevice
numpy
scipy
pydub
gtts
groq
langchain-groq
langchain-core
python-dotenv
openai-whisper
